# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Akmatori API
  description: |
    AI-powered AIOps platform API. Manages skills, tools, incidents, alert sources,
    and system settings for automated incident investigation and remediation.
  version: 1.0.0
  contact:
    name: Akmatori
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors
...
    Integration:
      type: object
      description: |
        A configured connection to a messaging provider (Slack, Telegram, ...).
        Credentials are stored as a JSON blob whose shape is provider-specific.
      properties:
        id: {type: integer}
        uuid: {type: string, format: uuid}
        provider:
          type: string
          enum: [slack, telegram]
        name: {type: string}
        credentials:
          type: object
          additionalProperties: true
        enabled: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    IntegrationCreateRequest:
      type: object
      required: [provider, name]
      properties:
        provider:
          type: string
          enum: [slack, telegram]
        name: {type: string}
        credentials:
          type: object
          additionalProperties: true
        enabled: {type: boolean}
    IntegrationUpdateRequest:
      type: object
      description: Provider is immutable on update.
      properties:
        name: {type: string}
        credentials:
          type: object
          additionalProperties: true
        enabled: {type: boolean}
    Channel:
      type: object
      description: |
        A specific addressable destination within an Integration (a Slack
        channel, a Telegram chat). Capability flags determine which triggers
        can reference it.
      properties:
        id: {type: integer}
        uuid: {type: string, format: uuid}
        integration_id: {type: integer}
        external_id:
          type: string
          description: Provider-specific channel handle (e.g. Slack channel ID or name)
        display_name: {type: string}
        can_post: {type: boolean}
        can_listen: {type: boolean}
        is_default_post:
          type: boolean
          description: At most one is_default_post=true per provider is enforced.
        extraction_prompt:
          type: string
          description: Optional per-channel system prompt for inbound extraction
        process_human_messages: {type: boolean}
        enabled: {type: boolean}
        integration: {$ref: '#/components/schemas/Integration'}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    ChannelCreateRequest:
      type: object
      required: [integration_uuid, external_id]
      properties:
        integration_uuid: {type: string, format: uuid}
        external_id: {type: string}
        display_name: {type: string}
        can_post: {type: boolean}
        can_listen: {type: boolean}
        is_default_post: {type: boolean}
        extraction_prompt: {type: string}
        process_human_messages: {type: boolean}
        enabled: {type: boolean}
    ChannelUpdateRequest:
      type: object
      description: Partial patch; integration cannot be changed after creation.
      properties:
        external_id: {type: string}
        display_name: {type: string}
        can_post: {type: boolean}
        can_listen: {type: boolean}
        is_default_post: {type: boolean}
        extraction_prompt: {type: string}
        process_human_messages: {type: boolean}
        enabled: {type: boolean}
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /integrations:
    get:
      summary: List messaging provider integrations
      operationId: listIntegrations
      tags: [Integrations]
      responses:
        '200':
          description: Configured integrations
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Integration'}
        '503':
          description: Channel service is not configured
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    post:
      summary: Create a new integration
      operationId: createIntegration
      tags: [Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IntegrationCreateRequest'}
      responses:
        '201':
          description: Integration created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Integration'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Conflict (duplicate name or invariant violation)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /integrations/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      summary: Get integration by UUID
      operationId: getIntegration
      tags: [Integrations]
      responses:
        '200':
          description: Integration details
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Integration'}
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update integration (provider is immutable)
      operationId: updateIntegration
      tags: [Integrations]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/IntegrationUpdateRequest'}
      responses:
        '200':
          description: Updated integration
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Integration'}
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete integration (cascades channels)
      operationId: deleteIntegration
      tags: [Integrations]
      responses:
        '204':
          description: Integration deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /channels:
    get:
      summary: List channels (optionally filtered)
      operationId: listChannels
      tags: [Channels]
      parameters:
        - name: integration_uuid
          in: query
          schema: {type: string, format: uuid}
          description: Filter to channels belonging to this integration
        - name: can_post
          in: query
          schema: {type: boolean}
          description: Filter to channels whose can_post flag matches
        - name: can_listen
          in: query
          schema: {type: boolean}
          description: Filter to channels whose can_listen flag matches
      responses:
        '200':
          description: Channels matching the filter
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Channel'}
        '503':
          description: Channel service is not configured
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    post:
      summary: Create a new channel
      operationId: createChannel
      tags: [Channels]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChannelCreateRequest'}
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Channel'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Integration not found
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
        '409':
          description: Default-post invariant violated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /channels/{uuid}:
    parameters:
      - name: uuid
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      summary: Get channel by UUID
      operationId: getChannel
      tags: [Channels]
      responses:
        '200':
          description: Channel details
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Channel'}
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update channel
      operationId: updateChannel
      tags: [Channels]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChannelUpdateRequest'}
      responses:
        '200':
          description: Updated channel
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Channel'}
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Default-post invariant violated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
    delete:
      summary: Delete channel
      operationId: deleteChannel
      tags: [Channels]
      responses:
        '204':
          description: Channel deleted
        '404':
          $ref: '#/components/responses/NotFound'
...